refactor: transition to /lib for airtable and slack and add module aliases
