+++ package.json
@@ -20,6 +20,12 @@
   "engines": {
     "node": "12.16.x"
   },
+  "_moduleAliases": {
+    "~airtable": "src/lib/airtable",
+    "~slack": "src/lib/slack",
+    "~assets": "src/lib/assets",
+    "~constants": "src/lib/constants.js"
+  },
   "proxy": "http://localhost:5000",
   "dependencies": {
     "@googlemaps/google-maps-services-js": "2.4.2",
@@ -43,6 +49,7 @@
     "lodash": "^4.17.15",
     "mapbox-gl": "^1.9.1",
     "mocha": "^7.1.1",
+    "module-alias": "2.2.2",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
     "react-helmet": "^5.2.1",
@@ -57,6 +64,7 @@
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-prettier": "^6.5.0",
     "eslint-config-react-app": "^5.0.2",
+    "eslint-import-resolver-alias": "1.1.2",
     "eslint-import-resolver-custom-alias": "^1.2.0",
     "eslint-plugin-flowtype": "4.7.0",
     "eslint-plugin-import": "2.20.2",
