+++ src/api/neighborhood-finder/update-request.js
@@ -1,4 +1,4 @@
-const { updateRequestByCode, findRequestByCode } = require("../../airtable.js");
+const { updateRequestByCode, findRequestByCode } = require("~airtable/bases");
 
 const NEIGHBORHOOD_AREA_UNAVAILABLE_OPTION = "Other - not Crown Heights";
 
@@ -38,5 +38,5 @@ exports.neighborhoodFinderUpdateRequestHandler = async (req, res) => {
     });
   }
 
-  res.send({ success: true });
+  return res.send({ success: true });
 };
