+++ src/lib/airtable/bases.js
@@ -164,7 +164,6 @@ exports.findVolunteerById = async id => {
 };
 
 exports.airbase = base;
-exports.UPDATE_BATCH_SIZE = 10;
 exports.SENSITIVE_FIELDS = [
   "Phone",
   "Email Address",
