+++ src/lib/airtable/constants.js
@@ -1,0 +1,1 @@
+exports.UPDATE_BATCH_SIZE = 10;
