+++ src/lib/assets/README.md
