+++ src/workers/airtable-sync/worker.js
@@ -1,6 +1,7 @@
 const ChangeDetector = require("airtable-change-detector");
+const { airbase, SENSITIVE_FIELDS } = require("~airtable/bases");
+const { UPDATE_BATCH_SIZE } = require("~airtable/constants");
 const updateMessageContent = require("./actions/updateMessageContent");
-const { SENSITIVE_FIELDS, airbase } = require("../../airtable");
 
 const defaultInterval = 5000;
